Fix 64-bit builds under Windows.
It's not entirely clear why wxWidgets chooses `_x64` as the path suffix,
or why we used `_amd64` before. Alas, with Visual C++ 14.0 (2015) `_x64` is
used by wxWidgets, so this is what we use as well.

+++ CaWE/MapEditor/DialogEditSurfaceProps.cpp
@@ -482,7 +482,7 @@ void EditSurfacePropsDialogT::ApplyClick(ViewWindow3DT& ViewWin3D, MapElementT* 
 {
     if (Object==NULL) return;
 
-    const ApplyModeT ApplyMode = (ApplyModeT)(unsigned long)ChoiceRightMBMode->GetClientData(ChoiceRightMBMode->GetSelection());
+    const ApplyModeT ApplyMode = (ApplyModeT)(uintptr_t)ChoiceRightMBMode->GetClientData(ChoiceRightMBMode->GetSelection());
 
     MapBrushT*       Brush  =dynamic_cast<MapBrushT*>(Object);
     MapBezierPatchT* Patch  =dynamic_cast<MapBezierPatchT*>(Object);
